Moved some stuff from controller into view.
