+++ class/Controller/RoleController.php
@@ -27,14 +27,18 @@ abstract class RoleController
 {
 
     protected $factory;
+    protected $view;
     protected $role;
     protected $id;
 
     abstract protected function loadFactory();
 
+    abstract protected function loadView();
+
     public function __construct($role)
     {
         $this->loadFactory();
+        $this->loadView();
         $this->role = $role;
     }
 
@@ -108,7 +112,7 @@ abstract class RoleController
         } else {
             $method_name = $command . 'PutCommand';
         }
-        
+
         if (!method_exists($this, $method_name)) {
             throw new BadCommand($method_name);
         }
@@ -131,6 +135,8 @@ abstract class RoleController
             if ($this->id && method_exists($this, 'viewHtmlCommand')) {
                 $method_name = 'viewHtmlCommand';
             } else {
+                //echo(method_exists($this, 'viewHtmlCommand'));
+                var_dump($this->id);
                 throw new BadCommand($method_name);
             }
         }
